+++ src/emu/cpu/pic16c5x/16c5xdsm.c
@@ -137,7 +137,7 @@ static void InitDasm16C5x(void)
 			fatalerror("not enough bits in encoding '%s %s' %d",
 				ops[0],ops[1],bit);
 		}
-		while (isspace(*p)) p++;
+		while (isspace((UINT8)*p)) p++;
 		if (*p) Op[i].extcode = *p;
 		Op[i].bits = bits;
 		Op[i].mask = mask;
